Reestructure cf table generation
This reestructure allows respectin g sorting criteria preference (i.e.
inequalities come first, and then query sorting is added if compatible).

Also, now those columns not present in the projection, but required
(mainly because they belong to the primary key, for one reason or
another) are also added to the resulting table